Changes to Card. (#29)
* Added social urls variables, they were not there.
Social's ul will now only display if either of the social url's is present. Did so to avoid an empty space on Card.
Changed 'overview' to use conditional in order to avoid empty paragraph being displayed.

* Added social urls variables, they were not there.
Social's ul will now only display if any of the social url's is present. Did so to avoid an empty space on Card.
Changed 'overview' to use conditional && in order to avoid empty paragraph being displayed.