Performance and readability in CardGrid
The mapStateToProps function in the CardGrid component was using a
double for-loop to filter valid results. This fixes that issue by using
a Set, which is O(1) time complexity, so the overall time complexity of
the loop is now O(n) rather than O(n^2).

Additionally, readability is improved by using less code with more
readable apis like map and filter instead of vanilla for loops.